AI-agents/requirements.txt
pydantic
aiohttp
python-dotenv
openai
